Add more download badges (#421)
- Adds more badges on the readme to "get it on" for github and
obtainium.
- Adds the hash of the signed certificate of each version on the readme.
- Adds the hash of the signed certificate for the github version on the
release note.

[Here how it
looks](https://github.com/PabloOQ/UrlChecker/tree/get-it-on#urlcheck)

The github badge simply redirects to the latest release.

I was going to explain what Obtainium does but [this
page](https://obtainium.imranr.dev/) explains it perfectly.
The hashes are added because Obtainium has compatibility with
[AppVerifier](https://github.com/soupslurpr/AppVerifier), which is an
app to check if the hash of the signed certificate is correct.

The obtainium badge sends you to a page which, on mobile, will then
redirect you to the obtainium app. The reason can be found
[here](https://github.com/ImranR98/apps.obtainium.imranr.dev?tab=readme-ov-file#url-redirection),
this repo also explains the 2 versions for the obtainium links `/add`
and `/app`. Because currently there is no complexity in the releases
version page (just one apk file), the `/add` link works out of the box,
only need to do a few confirmation taps, that's why I just exported the
resulting configuration and added it as an `/app` link, to save a few
taps. However the configuration URL is more cluttered.

<details>
<summary>see</summary>

---

http://apps.obtainium.imranr.dev/redirect.html?r=obtainium://add/https://github.com/TrianguloY/URLCheck

vs


https://apps.obtainium.imranr.dev/redirect?r=obtainium://app/%7B%22id%22%3A%22com.trianguloy.urlchecker%22%2C%22url%22%3A%22https%3A%2F%2Fgithub.com%2FTrianguloY%2FURLCheck%22%2C%22author%22%3A%22TrianguloY%22%2C%22name%22%3A%22URLCheck%22%2C%22preferredApkIndex%22%3A0%2C%22additionalSettings%22%3A%22%7B%5C%22includePrereleases%5C%22%3Afalse%2C%5C%22fallbackToOlderReleases%5C%22%3Atrue%2C%5C%22filterReleaseTitlesByRegEx%5C%22%3A%5C%22%5C%22%2C%5C%22filterReleaseNotesByRegEx%5C%22%3A%5C%22%5C%22%2C%5C%22verifyLatestTag%5C%22%3Afalse%2C%5C%22dontSortReleasesList%5C%22%3Afalse%2C%5C%22useLatestAssetDateAsReleaseDate%5C%22%3Afalse%2C%5C%22releaseTitleAsVersion%5C%22%3Afalse%2C%5C%22trackOnly%5C%22%3Afalse%2C%5C%22versionExtractionRegEx%5C%22%3A%5C%22%5C%22%2C%5C%22matchGroupToUse%5C%22%3A%5C%22%5C%22%2C%5C%22versionDetection%5C%22%3Atrue%2C%5C%22releaseDateAsVersion%5C%22%3Afalse%2C%5C%22useVersionCodeAsOSVersion%5C%22%3Afalse%2C%5C%22apkFilterRegEx%5C%22%3A%5C%22%5C%22%2C%5C%22invertAPKFilter%5C%22%3Afalse%2C%5C%22autoApkFilterByArch%5C%22%3Atrue%2C%5C%22appName%5C%22%3A%5C%22%5C%22%2C%5C%22shizukuPretendToBeGooglePlay%5C%22%3Afalse%2C%5C%22allowInsecure%5C%22%3Afalse%2C%5C%22exemptFromBackgroundUpdates%5C%22%3Afalse%2C%5C%22skipUpdateNotifications%5C%22%3Afalse%2C%5C%22about%5C%22%3A%5C%22%5C%22%2C%5C%22refreshBeforeDownload%5C%22%3Afalse%7D%22%2C%22overrideSource%22%3Anull%7D

---
</details>

More info about Obtainium:
- https://github.com/ImranR98/Obtainium
- https://wiki.obtainium.imranr.dev/


Also, IzzyOnDroid is a popular F-Droid alternative. Do you mind if I ask
there to have URLChecker added? Although, this is more of a courtesy, as
technically anyone can ask, this is because Izzy uses the releases from
github and doesn't need anything else (in fact izzy actively adds apps
they like). If that's the case then, I would add a commit to this PR
with the corresponding badge.
[More about IzzyOnDroid.](https://apt.izzysoft.de/fdroid/index/info)
[Badge source (for
later).](https://gitlab.com/IzzyOnDroid/repo/#if-my-app-is-available-here-do-you-have-a-badge-i-can-use-to-link-to-it)

---------

Co-authored-by: TrianguloY <correo--correo@hotmail.com>